+++ parser/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add error types `RecursiveNewtype` and `InvalidKeyType`.
 - Add warning type `NonCamelCaseNewtype`.
 - Support newtypes as the key type in `set<T>` and `map<K -> T>`.
+- Add warning `ReservedSchemaName`.
 
 ### Changed